Moved some folders around
#1251 moved some folders around
