+++ packages/docs/src/components/DocsSubtitle.vue
@@ -13,7 +13,7 @@
 import { TranslationString } from '../components/DocsApi/ManualApiOptions'
 import { Options, Vue, mixins, prop } from 'vue-class-component'
 import DocsAnchor from './DocsAnchor.vue'
-import MarkdownView from '../utilities/markdown-view/MarkdownView.vue'
+import MarkdownView from './markdown-view/MarkdownView.vue'
 
 class Props {
   text = prop<TranslationString>({ type: String, required: true })
