+++ packages/docs/src/components/header/components/HeaderSelector.vue
@@ -17,8 +17,8 @@
 
 <script lang="ts">
 
-import VaIconMenu from '../../../iconset/VaIconMenu.vue'
-import VaIconMenuCollapsed from '../../../iconset/VaIconMenuCollapsed.vue'
+import VaIconMenu from '../../iconset/VaIconMenu.vue'
+import VaIconMenuCollapsed from '../../iconset/VaIconMenuCollapsed.vue'
 import { getColors } from '../../../../../ui/src/main'
 
 export default {
