+++ packages/docs/src/components/iconset/VaIconMenuCollapsed.vue
