+++ packages/docs/src/components/markdown-view/set-class-attribute-to-external-links.ts
