+++ packages/docs/src/components/markdown-view/set-origin-location-to-relative-links.ts
